+++ logmanager/option.go
@@ -47,16 +47,6 @@ func WithTraceIDHeaderKey(key string) Option {
 	}
 }
 
-// WithMaskConfigs configures the application with a list of MaskConfigs for masking sensitive data in logs.
-// Deprecated: Use WithMaskingConfig with JSONPath support instead.
-func WithMaskConfigs(maskConfigs MaskConfigs) Option {
-	return func(app *Application) {
-		if len(maskConfigs) > 0 {
-			app.maskConfigs = maskConfigs
-		}
-	}
-}
-
 // WithMaskingConfig configures the application with a list of MaskingConfig for masking sensitive data in logs using JSONPath.
 func WithMaskingConfig(maskingConfigs []MaskingConfig) Option {
 	return func(app *Application) {